IDL2K better dictionary support: generate interfaces with builder function suppress inline warning, replace Long with Int (cherry picked from commit 342f35f)

+++ libraries/tools/idl2k/src/main/kotlin/typeMappings.kt
@@ -20,8 +20,8 @@ import java.util.*
 
 private val typeMapper = mapOf(
         "unsignedlong" to SimpleType("Int", false),
-        "unsignedlonglong" to SimpleType("Long", false),
-        "longlong" to SimpleType("Long", false),
+        "unsignedlonglong" to SimpleType("Int", false),
+        "longlong" to SimpleType("Int", false),
         "unsignedshort" to SimpleType("Short", false),
         "unsignedbyte" to SimpleType("Byte", false),
         "octet" to SimpleType("Byte", false),
